+++ src/Core/Parser/Entity/CollectionLogOperation/OperationsCollection.php
@@ -98,7 +98,7 @@ final class OperationsCollection implements \IteratorAggregate
                 $entityName = $entityName && $entity?->isEntityDataCanBeLoaded() ? $entityName : null;
                 if ($operation->getEntityName() !== $entityName) {
                     return true;
-                } 
+                }
                 if ($entity?->isEntityCacheOutdated() && $entity?->isEntityDataCanBeLoaded()) {
                     return true;
                 }